Rename classes and improve documentation
