+++ src/main/java/duke/Duke.java
@@ -2,24 +2,15 @@ package duke;
 
 import duke.views.cli.Cli;
 import duke.views.gui.Launcher;
-import duke.views.strategies.MultiType;
+import duke.views.strategies.Parser;
 import javafx.application.Application;
 
 public class Duke {
     /**
-     * Initial method to test that the main method can run properly.
-     */
-    public static void test() {
-        String logo = " ____        _        \n" + "|  _ \\ _   _| | _____ \n" + "| | | | | | | |/ / _ \\\n"
-                + "| |_| | |_| |   <  __/\n" + "|____/ \\__,_|_|\\_\\___|\n";
-        System.out.println("Hello from\n" + logo);
-    }
-
-    /**
      * Runs the CLI application.
      */
     public static void runCli() {
-        Cli chatbot = new Cli(new MultiType());
+        Cli chatbot = new Cli(new Parser());
         chatbot.main();
     }
