+++ src/main/java/duke/views/gui/Launcher.java
@@ -2,8 +2,8 @@ package duke.views.gui;
 
 import java.io.IOException;
 
-import duke.views.strategies.MultiType;
 import duke.views.gui.controllers.MainWindow;
+import duke.views.strategies.Parser;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -16,7 +16,7 @@ public class Launcher extends Application {
     @Override
     public void start(Stage stage) {
         try {
-            this.gui = new Gui(new MultiType());
+            this.gui = new Gui(new Parser());
             this.gui.initCallbacks();
             FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/MainWindow.fxml"));
             AnchorPane ap = fxmlLoader.load();
